// frontend/src/types/random.d.ts
export interface IDelivery {
    _id: string;
    dietId: {
        _id: string;
        patientId: {
            name: string;
            roomNumber: string;
            bedNumber: string;
        };
        meals: {
            morning: {
                _id: string;
                items: any[];
                status: string;
            };
            evening: {
                _id: string;
                items: any[];
                status: string;
            };
            night: {
                _id: string;
                items: any[];
                status: string;
            };
        };
    };
    assignedTo?: {
        name?: string;
        email?: string;
        contact?: string;
    };
    mealTime: "morning" | "evening" | "night";
    status: "assigned" | "in-progress" | "delivered";
    deliveredAt?: Date;
    notes?: string;
}